+++ src/app/api/total-supply/route.ts
@@ -6,6 +6,7 @@ import LSTService from "@/services/lst";
 
 export const revalidate = 120;
 
+//TODO: remove if not needed
 export async function GET(_req: Request) {
   const provider = getProvider();